Initial commit
Initial commit for the REDCap Branching Logic parser. This parser uses
the PyParsing library to parse and evaluate REDCap's branching logic.
It parses the branching logic string of a REDCap field by converting
the string into an abstract syntax tree (AST) containing the fields to
be looked up, and then looks up and substitutes the fields in the AST
for their boolean values. Finally, the boolean expression is evaluated,
and the output of the branching logic is returned.

Right now for the initial commit, the parser works. However, further
testing is still needed. In particular, further unit test coverage
is required before this project can be deployed in a production use.